fix functions and refactoring functions
