+++ project/database/models.py
@@ -184,9 +184,7 @@ class QhawaxInstallationHistory(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     lat = db.Column(db.Float)
     lon = db.Column(db.Float)
-    instalation_date = db.Column(db.DateTime, nullable=False)
     installation_date_zone = db.Column(db.DateTime, nullable=False)
-    end_date = db.Column(db.DateTime, nullable=False)
     end_date_zone = db.Column(db.DateTime, nullable=False)
     link_report = db.Column(db.String(500), nullable=False, unique=True)
     observations = db.Column(db.String(300), nullable=False, unique=True)
