+++ project/main/business/gas_sensor.py
@@ -1,11 +1,8 @@
 from flask import jsonify, make_response, request
-import datetime
-import dateutil.parser
-import dateutil.tz
-
-from project import app, db, socketio
 from project.database.models import Qhawax, GasSensor
-import project.main.business.business_helper as helper
+import project.main.business.get_business_helper as get_business_helper
+import project.main.business.post_business_helper as post_business_helper
+from project import app
 
 @app.route('/api/request_offsets/', methods=['GET'])
 def requestOffsets():
@@ -18,7 +15,7 @@ def requestOffsets():
     """
     qhawax_name = request.args.get('ID')
     try:
-        offsets = helper.getOffsetsFromProductID(qhawax_name)
+        offsets = get_business_helper.getOffsetsFromProductID(qhawax_name)
         return make_response(jsonify(offsets), 200)
     except TypeError as e:
         json_message = jsonify({'error': '\'%s\'' % (e)})
@@ -35,7 +32,7 @@ def requestControlledOffsets():
     """
     qhawax_name = request.args.get('ID')
     try:
-        controlled_offsets = helper.getControlledOffsetsFromProductID(qhawax_name)
+        controlled_offsets = get_business_helper.getControlledOffsetsFromProductID(qhawax_name)
         return make_response(jsonify(controlled_offsets), 200)
     except TypeError as e:
         json_message = jsonify({'error': '\'%s\'' % (e)})
@@ -52,7 +49,7 @@ def requestNonControlledOffsets():
     """
     qhawax_name = request.args.get('ID')
     try:
-        non_controlled_offsets = helper.getNonControlledOffsetsFromProductID(qhawax_name)
+        non_controlled_offsets = get_business_helper.getNonControlledOffsetsFromProductID(qhawax_name)
         print(non_controlled_offsets)
         return make_response(jsonify(non_controlled_offsets), 200)
     except TypeError as e:
@@ -73,14 +70,14 @@ def saveOffsets():
     """
     req_json = request.get_json()
     try:
-        qhawax_id = str(req_json['product_id']).strip()
+        qhawax_name = str(req_json['product_id']).strip()
         offsets = req_json['offsets']
         try:
-            helper.updateOffsetsFromProductID(qhawax_id, offsets)
+            post_business_helper.updateOffsetsFromProductID(qhawax_name, offsets)
             description="Se actualizaron constantes offsets"
             observation_type="Interna"
             person_in_charge = req_json['person_in_charge']
-            helper.writeBitacora(qhawax_id,observation_type,description,person_in_charge)
+            post_business_helper.writeBitacora(qhawax_name,observation_type,description,person_in_charge)
             return make_response('Success', 200)
         except TypeError as e:
             json_message = jsonify({'error': '\'%s\'' % (e)})
@@ -103,17 +100,17 @@ def saveControlledOffsets():
     """
     req_json = request.get_json()
     try:
-        qhawax_id = str(req_json['product_id']).strip()
+        qhawax_name = str(req_json['product_id']).strip()
         controlled_offsets = req_json['controlled_offsets']
     except KeyError as e:
         json_message = jsonify({'error': 'Parameter \'%s\' is missing in JSON object' % (e)})
         return make_response(json_message, 400)
 
-    helper.updateControlledOffsetsFromProductID(qhawax_id, controlled_offsets)
+    post_business_helper.updateControlledOffsetsFromProductID(qhawax_name, controlled_offsets)
     description="Se actualizaron constantes controladas"
     observation_type="Interna"
     person_in_charge = req_json['person_in_charge']
-    helper.writeBitacora(qhawax_id,observation_type,description,person_in_charge)
+    post_business_helper.writeBitacora(qhawax_name,observation_type,description,person_in_charge)
     return make_response('Success', 200)
 
 @app.route('/api/save_non_controlled_offsets/', methods=['POST'])
@@ -130,16 +127,16 @@ def saveNonControlledOffsets():
     """
     req_json = request.get_json()
     try:
-        qhawax_id = str(req_json['product_id']).strip()
+        qhawax_name = str(req_json['product_id']).strip()
         non_controlled_offsets = req_json['non_controlled_offsets']
     except KeyError as e:
         json_message = jsonify({'error': 'Parameter \'%s\' is missing in JSON object' % (e)})
         return make_response(json_message, 400)
 
-    helper.updateNonControlledOffsetsFromProductID(qhawax_id, non_controlled_offsets)
+    post_business_helper.updateNonControlledOffsetsFromProductID(qhawax_name, non_controlled_offsets)
     description="Se actualizaron constantes no controladas"
     observation_type="Interna"
     person_in_charge = req_json['person_in_charge']
-    helper.writeBitacora(qhawax_id,observation_type,description,person_in_charge)
+    post_business_helper.writeBitacora(qhawax_name,observation_type,description,person_in_charge)
     return make_response('Success', 200)
